fix: PacketType을 byte로 읽지 않고 int로 읽어 생기던 문제 수정

+++ Common/Serialization/BinaryReaderExtensions.cs
@@ -10,7 +10,7 @@ namespace PeachGame.Common.Serialization {
 		}
 
 		public static IPacket ReadPacket(this BinaryReader reader) {
-			var packetId = reader.ReadInt32();
+			var packetId = reader.ReadByte();
 			var packetType = (PacketType)packetId;
 			return packetType.CreatePacket(reader);
 		}
